+++ packages/generator/src/oot/save.c
@@ -1,4 +1,5 @@
 #include <combo.h>
+#include <combo/age.h>
 #include <combo/item.h>
 #include <combo/dungeon.h>
 #include <combo/dma.h>
@@ -15,7 +16,11 @@ static void fixSpawn(void)
     gSaveContext.respawnFlag = 0;
     gSaveContext.respawn[RESPAWN_MODE_DOWN].entrance = 0xffff;
 
-    if (gSharedCustomSave.respawn[CUSTOM_RESPAWN_MODE_DUNGEON_ENTRANCE].playerParams && !gComboCtx.valid && gPlay)
+    /* Don't override anything if it's not a cold start */
+    if (gComboCtx.valid)
+        return;
+
+    if (gSharedCustomSave.respawn[CUSTOM_RESPAWN_MODE_DUNGEON_ENTRANCE].playerParams && gPlay)
     {
         if (gSharedCustomSave.respawn[CUSTOM_RESPAWN_MODE_DUNGEON_ENTRANCE].data & 0x80)
         {
@@ -33,11 +38,17 @@ static void fixSpawn(void)
         }
     }
 
-    /* If the player saved in link's house, and it's not ER, honor that */
-    if (gSave.info.sceneId == SCE_OOT_LINK_HOUSE && !Config_Flag(CFG_ER_ANY))
+    /* If the player saved in link's house, honor that */
+    if (gSave.info.sceneId == SCE_OOT_LINK_HOUSE)
     {
         gSave.entrance = ENTR_OOT_SPAWN_CHILD;
         return;
+    }
+
+    /* If the player hasn't reached ToT age swap and is supposed to reach their default spawn, ensure they're set to their default age */
+    if (!gCustomSave.hasTimeTraveledAtTemple)
+    {
+        Age_SetOot(NULL, Age_GetStarting());
     }
 
     entrance = gSave.age == AGE_CHILD ? ENTR_OOT_SPAWN_CHILD : ENTR_OOT_SPAWN_ADULT;